combine load and save file
